+++ src/srv/events/mailing_created.py
@@ -28,7 +28,7 @@ from databases import Database
 from eljson.json import Json
 from pyeo import elegant
 
-from app_types.listable import FkAsyncListable
+from app_types.fk_async_listable import FkAsyncListable
 from app_types.logger import LogSink
 from app_types.update import FkUpdate
 from exceptions.internal_exceptions import TelegramIntegrationsError, UnreacheableError
