+++ src/srv/events/morning_content_published.py
@@ -31,7 +31,7 @@ from eljson.json import Json
 from jinja2 import Template
 from pyeo import elegant
 
-from app_types.listable import FkAsyncListable
+from app_types.fk_async_listable import FkAsyncListable
 from app_types.logger import LogSink
 from app_types.update import FkUpdate
 from exceptions.internal_exceptions import TelegramIntegrationsError
