+++ src/srv/events/prayers_mailing.py
@@ -33,7 +33,7 @@ from eljson.json import Json
 from pyeo import elegant
 from redis.asyncio import Redis
 
-from app_types.listable import FkAsyncListable
+from app_types.fk_async_listable import FkAsyncListable
 from app_types.logger import LogSink
 from app_types.update import FkUpdate
 from exceptions.internal_exceptions import TelegramIntegrationsError
